+++ frontend/catalyst/device/verification.py
@@ -30,7 +30,7 @@ from pennylane.measurements import (
     VnEntropyMP,
 )
 from pennylane.measurements.shots import Shots
-from pennylane.operation import Operation, Tensor
+from pennylane.operation import Operation, StatePrepBase, Tensor
 from pennylane.ops import (
     Adjoint,
     CompositeOp,
@@ -209,7 +209,13 @@ def verify_operations(tape: QuantumTape, grad_method, qjit_device):
         # is handled in _inv_op_checker and _ctrl_op_checker.
         # Specialed control op classes (e.g. CRZ) should be checked directly though, which is why we
         # can't use isinstance(op, Controlled).
-        if type(op) in (Controlled, ControlledOp) or isinstance(op, Adjoint):
+        if type(op) in (Controlled, ControlledOp) or isinstance(op, (Adjoint)):
+            pass
+        # Don't check StatePrep since StatePrep is not in the list of device capabilities.
+        # It is only valid when the TOML file has the initial_state_prep_flag.
+        elif (
+            isinstance(op, StatePrepBase) and qjit_device.qjit_capabilities.initial_state_prep_flag
+        ):
             pass
         elif not qjit_device.qjit_capabilities.native_ops.get(op.name):
             raise CompileError(