+++ mlir/include/Quantum/IR/QuantumOps.h
@@ -23,6 +23,7 @@
 #include "mlir/Interfaces/SideEffectInterfaces.h"
 #include "mlir/Support/LogicalResult.h"
 #include "llvm/ADT/StringRef.h"
+#include "llvm/Support/ErrorHandling.h"
 #include <optional>
 
 #include "Quantum/IR/QuantumInterfaces.h"