+++ runtime/lib/backend/lightning/lightning_dynamic/LightningSimulator.cpp
@@ -104,6 +104,21 @@ void LightningSimulator::PrintState()
     cout << state[idx] << "]" << endl;
 }
 
+void LightningSimulator::SetState(DataView<std::complex<double>, 1> &data,
+                                  std::vector<QubitIdType> &wires)
+{
+    std::vector<std::complex<double>> data_vector(data.begin(), data.end());
+    std::vector<std::size_t> wires_size_t(wires.begin(), wires.end());
+    this->device_sv->setStateVector(data_vector, wires_size_t);
+}
+
+void LightningSimulator::SetBasisState(DataView<int8_t, 1> &data, std::vector<QubitIdType> &wires)
+{
+    std::vector<std::size_t> data_vector(data.begin(), data.end());
+    std::vector<std::size_t> wires_size_t(wires.begin(), wires.end());
+    this->device_sv->setBasisState(data_vector, wires_size_t);
+}
+
 auto LightningSimulator::Zero() const -> Result
 {
     return const_cast<Result>(&GLOBAL_RESULT_FALSE_CONST);